// src/platform/plugins/shared/files/common/types.d.ts
export interface FileKind extends FileKindBase {
    /**
     * Max file contents size, in bytes. Can be customized per file using the
     * {@link FileJSON} object. This is enforced on the server-side as well as
     * in the upload React component.
     *
     * @default 4MiB
     */
    maxSizeBytes?: number | ((file: FileJSON) => number);
    /**
     * Blob store specific settings that enable configuration of storage
     * details.
     */
    blobStoreSettings?: BlobStorageSettings;
    /**
     * Specify which HTTP routes to create for the file kind.
     *
     * You can always create your own HTTP routes for working with files but
     * this interface allows you to expose basic CRUD operations, upload, download
     * and sharing of files over a RESTful-like interface.
     *
     * @note The public {@link FileClient} uses these endpoints.
     */
    http: {
        /**
         * Expose file creation (and upload) over HTTP.
         */
        create?: HttpEndpointDefinition;
        /**
         * Expose file updates over HTTP.
         */
        update?: HttpEndpointDefinition;
        /**
         * Expose file deletion over HTTP.
         */
        delete?: HttpEndpointDefinition;
        /**
         * Expose "get by ID" functionality over HTTP.
         */
        getById?: HttpEndpointDefinition;
        /**
         * Expose the ability to list all files of this kind over HTTP.
         */
        list?: HttpEndpointDefinition;
        /**
         * Expose the ability to download a file's contents over HTTP.
         */
        download?: HttpEndpointDefinition;
        /**
         * Expose file share functionality over HTTP.
         */
        share?: HttpEndpointDefinition;
    };
    /**
     * A list of hashes to compute for this file kind. The hashes will be computed
     * during the file upload process and stored in the file metadata.
     */
    hashes?: SupportedFileHashAlgorithm[];
}
/** Definition for an endpoint that the File's service will generate */
interface HttpEndpointDefinition {
    /**
     * Specify the tags for this endpoint.
     *
     * @example
     * // This will enable access control to this endpoint for users that can access "myApp" only.
     * { requiredPrivileges: ['myApp'] }
     *
     */
    requiredPrivileges: string[];
}
// ...
export interface BlobStorageSettings {
    /**
     * Single index that supports up to 50GB of blobs
     */
    [ES_FIXED_SIZE_INDEX_BLOB_STORE]?: {
        index: string;
    };
}